AUDIT: autogenerated Localizable.strings
This is getting closer, but I'm still seeing these two oddities:

Some existing translations not found, maybe because of the lack of a
space around the `=` in the original? Honestly I could skip this one
assuming that going forward errors like this will be rare because we
won't be manually generating the strings file.

```
-/* Activity indicator title, shown upon returning to the device manager, until you complete the provisioning process on desktop */
-WAITING_TO_COMPLETE_DEVICE_LINK_TEXT="Complete setup on Signal Desktop.";
+/* Activity indicator title, shown upon returning to the device manager, until you complete the provisioning process on desktop */
+"WAITING_TO_COMPLETE_DEVICE_LINK_TEXT" = "WAITING_TO_COMPLETE_DEVICE_LINK_TEXT";
```

Though this one is more severe, seems like multi-line strings aren't
correctly captured:

```
+/* Paragraph(s) shown alongside keying material when verifying privacy with {{contact name}} */
+"PRIVACY_VERIFICATION_INSTRUCTIONS" = ";
-/* Paragraph(s) shown alongside keying material when verifying privacy with {{contact name}} */
-"PRIVACY_VERIFICATION_INSTRUCTIONS" = "If you wish to verify the security of your end-to-end encryption with %@, compare the numbers above with the numbers on their device.

```